backend/requirements.txt
langgraph==0.2.34
langchain-core==0.3.15
groq==0.11.0
gtts==2.5.3
google-api-python-client==2.147.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
httpx==0.27.2
pydub
elevenlabs
pillow 
requests
replicate
stability-sdk